Verilog: extract scope data structure from parser
This extracts the scope data structure from the parser class into a separate
file, as the scopes need to be preserved until all Verilog parsing is
finished.
