+++ src/verilog/verilog_parser.cpp
@@ -53,32 +53,3 @@ bool parse_verilog_file(const std::string &filename, verilog_standardt standard)
 
   return verilog_parser.parse();
 }
-
-/*******************************************************************\
-
-Function: verilog_parsert::lookup
-
-  Inputs:
-
- Outputs:
-
- Purpose:
-
-\*******************************************************************/
-
-const verilog_parsert::scopet *verilog_parsert::lookup(irep_idt name) const
-{
-  // we start from the current scope, and walk upwards to the root
-  auto scope = current_scope;
-  while(scope != nullptr)
-  {
-    auto name_it = scope->scope_map.find(name);
-    if(name_it == scope->scope_map.end())
-      scope = scope->parent;
-    else
-      return &name_it->second; // found it
-  }
-
-  // not found, give up
-  return nullptr;
-}
